+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Ability to get component data from entity handle.
+
+### Fixed
+
+- Doesn't add duplicate components anymore.
+
 ## [0.1.0] - 2021-04-17
 
 ### Added